refactor(sales): padronizar tipos financeiros de decimal para long

+++ template/backend/src/Ambev.DeveloperEvaluation.Application/Sales/GetSale/GetSaleResult.cs
@@ -18,7 +18,7 @@ namespace Ambev.DeveloperEvaluation.Application.Sales.GetSale
         public string CustomerName { get; set; } = string.Empty;
         public Guid BranchId { get; set; }
         public string BranchName { get; set; } = string.Empty;
-        public decimal TotalAmount { get; set; }
+        public long TotalAmount { get; set; }
         public bool IsCancelled { get; set; }
         public List<GetSaleItemResult> Items { get; set; } = new();
     }
@@ -29,9 +29,9 @@ namespace Ambev.DeveloperEvaluation.Application.Sales.GetSale
         public Guid ProductId { get; set; }
         public string ProductName { get; set; } = string.Empty;
         public int Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
-        public decimal Discount { get; set; }
-        public decimal TotalAmount { get; set; }
+        public long UnitPrice { get; set; }
+        public long Discount { get; set; }
+        public long TotalAmount { get; set; }
         public bool IsCancelled { get; set; }
     }
 }
