+++ template/backend/src/Ambev.DeveloperEvaluation.WebApi/Features/Sales/GetSale/GetSaleResponse.cs
@@ -9,7 +9,7 @@
         public string CustomerName { get; set; } = string.Empty;
         public Guid BranchId { get; set; }
         public string BranchName { get; set; } = string.Empty;
-        public decimal TotalAmount { get; set; }
+        public long TotalAmount { get; set; }
         public bool IsCancelled { get; set; }
         public List<GetSaleItemResponse> Items { get; set; } = new();
     }
@@ -20,9 +20,9 @@
         public Guid ProductId { get; set; }
         public string ProductName { get; set; } = string.Empty;
         public int Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
-        public decimal Discount { get; set; }
-        public decimal TotalAmount { get; set; }
+        public long UnitPrice { get; set; }
+        public long Discount { get; set; }
+        public long TotalAmount { get; set; }
         public bool IsCancelled { get; set; }
     }
 }
