+++ template/backend/src/Ambev.DeveloperEvaluation.WebApi/Features/Sales/UpdateSale/UpdateSaleResponse.cs
@@ -4,7 +4,7 @@
     {
         public Guid Id { get; set; }
         public string SaleNumber { get; set; } = string.Empty;
-        public decimal TotalAmount { get; set; }
+        public long TotalAmount { get; set; }
         public bool IsCancelled { get; set; }
     }
 }
